[FIX] Pass to render into sdl

+++ src/Core/include.h
@@ -8,6 +8,6 @@
 using Vector2u = Vector2<int32>;
 using Vector2f = Vector2<float32>;
 
-constexpr float32 PI = 3.14159265358979323846;
+constexpr float64 PI = 3.14159265358979323846;
 constexpr float RAD_TO_DEG = 180.0f / PI;
 constexpr float DEG_TO_RAD = PI / 180.0f;
